huffman node files renamed
