+++ semester1/hw9/9.3/huffman.cpp
@@ -4,7 +4,7 @@
 #include <queue>
 #include <stack>
 #include "huffman.h"
-#include "node.h"
+#include "huffmanNode.h"
 #include "freqtable.h"
 
 int const alphabet = 256;
