+++ semester1/hw9/9.3/huffmanNode.cpp
@@ -1,4 +1,4 @@
-#include "node.h"
+#include "huffmanNode.h"
 
 bool isLeaf(HuffmanNode *node) {
     return (node->l == nullptr && node->r == nullptr);
