+++ semester1/hw9/9.3/list/list.h
@@ -1,4 +1,5 @@
 #pragma once
+#include "huffmanNode.h"
 
 struct ListNode;
